feat: merge monitoring branch to main (#51)
* feat: change name from anthos to northstar

* feat: add grafana and prometheus

* feat: update start and destroy scripts for new components

* fix: address bugs in prom and grafana deploy

* feat: add postgres monitoring containers to anthos

* fix: clean up logic for postgres monitor

* fix: minor comment update

* feat: update loadbalancer to be interactive and expose the gui

* fix: remove resource constraints from anthos

* fix: add database to postgres prom exporter connection strings

* feat: add service for loadgenerator

* feat: add test script to forward ports for testing

* feat: updated with new information on monitoring deployment

* fix: typo in configuration key

* fix: rename bank of anthos to bank of sirius (#46)

* chore: merge current monitoring changes back to feature branch (#49)

* docs: readme changes for modern apps context

* docs: fix image link

* feat: change name from anthos to northstar

* feat: add grafana and prometheus

* feat: update start and destroy scripts for new components

* fix: address bugs in prom and grafana deploy

* feat: add postgres monitoring containers to anthos

* fix: clean up logic for postgres monitor

* fix: minor comment update

* feat: update loadbalancer to be interactive and expose the gui

* fix: remove resource constraints from anthos

* fix: add database to postgres prom exporter connection strings

* feat: add service for loadgenerator

* feat: add test script to forward ports for testing

* feat: updated with new information on monitoring deployment

* fix: typo in configuration key

* fix: rename bank of anthos to bank of sirius

* feat: pull bank of sirius images from GH package registry

* feat: change test script to account for rename of anthos to sirius

* fix: remove unused code

* fix: add password prompt for grafana admin password in startup

* chore: reformat code to fit standards

* feat: add configuration options for kic helm chart

* feat: add configuration options for all helm chart deploys

* feat: update all manifests to use the project (bank of sirius) images

* docs: readme changes for modern apps context

* docs: fix image link

* feat: change name from anthos to northstar

* feat: add grafana and prometheus

* feat: update start and destroy scripts for new components

* fix: address bugs in prom and grafana deploy

* feat: add postgres monitoring containers to anthos

* fix: minor comment update

* feat: update loadbalancer to be interactive and expose the gui

* fix: rename bank of anthos to bank of sirius

* feat: pull bank of sirius images from GH package registry

* feat: change test script to account for rename of anthos to sirius

* fix: remove unused code

* fix: add password prompt for grafana admin password in startup

* chore: reformat code to fit standards

* feat: add configuration options for kic helm chart

* feat: add configuration options for all helm chart deploys

* feat: update all manifests to use the project (bank of sirius) images


* refactor: simplified destroy script

* feat: allow for custom config of VPC AZs

This change allows for users to configure what availability zones they
want to associate with the subnets on their VPC.

* chore: bump default helm chart version

* feat: verify aws creds in start and destroy scripts

* feat: verify k8s cluster has been stood up

* feat: improve logging when verifying aws creds

* feat: allow for the destruction of only k8s projects

* fix: select default region from AWS config rather than actual

* chore: disable self-signed cert warning for endpoint test

* fix: use aws and kubectl from path if available

* feat: fixes #33

This change allows a user to get NGINX Kubernetes Ingress Controller
directly from a container registry instead of having to build it from
source.

* fix: clarified region selection logic in startup script

* feat: add additional input validations

* fix: added missing properties

* docs: add additional image_name example

* refactor: move more functionality into base provider class

+++ pulumi/aws/destroy.sh
@@ -1,15 +1,15 @@
 #!/usr/bin/env bash
 
-set -o errexit   # abort on nonzero exit status
-set -o nounset   # abort on unbound variable
-set -o pipefail  # don't hide errors within pipes
+set -o errexit  # abort on nonzero exit status
+set -o nounset  # abort on unbound variable
+set -o pipefail # don't hide errors within pipes
 
 # Don't pollute console output with upgrade notifications
 export PULUMI_SKIP_UPDATE_CHECK=true
 # Run Pulumi non-interactively
 export PULUMI_SKIP_CONFIRMATIONS=true
 
-script_dir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+script_dir="$(cd "$(dirname "${BASH_SOURCE[0]}")" > /dev/null 2>&1 && pwd)"
 
 if ! command -v pulumi > /dev/null; then
   if [ -x "${script_dir}/venv/bin/pulumi" ]; then
@@ -17,17 +17,17 @@ if ! command -v pulumi > /dev/null; then
     export PATH="${script_dir}/venv/bin:$PATH"
 
     if ! command -v pulumi > /dev/null; then
-      >&2 echo "Pulumi must be installed to continue"
+      echo >&2 "Pulumi must be installed to continue"
       exit 1
     fi
   else
-    >&2 echo "Pulumi must be installed to continue"
+    echo >&2 "Pulumi must be installed to continue"
     exit 1
   fi
 fi
 
 if ! command -v python3 > /dev/null; then
-  >&2 echo "Python 3 must be installed to continue"
+  echo >&2 "Python 3 must be installed to continue"
   exit 1
 fi
 
@@ -37,11 +37,11 @@ if ! command -v node > /dev/null; then
     export PATH="${script_dir}/venv/bin:$PATH"
 
     if ! command -v node > /dev/null; then
-      >&2 echo "NodeJS must be installed to continue"
+      echo >&2 "NodeJS must be installed to continue"
       exit 1
     fi
   else
-    >&2 echo "NodeJS must be installed to continue"
+    echo >&2 "NodeJS must be installed to continue"
     exit 1
   fi
 fi
@@ -51,7 +51,7 @@ if ! pulumi whoami --non-interactive > /dev/null 2>&1; then
   pulumi login
 
   if ! pulumi whoami --non-interactive > /dev/null 2>&1; then
-    >&2 echo "Unable to login to Pulumi - exiting"
+    echo >&2 "Unable to login to Pulumi - exiting"
     exit 2
   fi
 fi
@@ -59,34 +59,62 @@ fi
 source "${script_dir}/config/environment"
 echo "Configuring all Pulumi projects to use the stack: ${PULUMI_STACK}"
 
-pulumi_args="--emoji --stack ${PULUMI_STACK}"
+function validate_aws_credentials() {
+  pulumi_aws_profile="$(pulumi --cwd "${script_dir}/vpc" config get aws:profile)"
+  if [ "${pulumi_aws_profile}" != "" ]; then
+    profile_arg="--profile ${pulumi_aws_profile}"
+  elif [[ -n "${AWS_PROFILE+x}" ]]; then
+    profile_arg="--profile ${AWS_PROFILE}"
+  else
+    profile_arg=""
+  fi
 
-cd "${script_dir}/anthos"
-pulumi ${pulumi_args} destroy
+  echo "Validating AWS credentials"
+  if ! "${script_dir}/venv/bin/aws" ${profile_arg} sts get-caller-identity > /dev/null; then
+    echo >&2 "AWS credentials have expired or are not valid"
+    exit 2
+  fi
+}
 
-cd "${script_dir}/certmgr"
-pulumi ${pulumi_args} destroy
+function destroy_project() {
+  local project_dir="${script_dir}/$1"
+  local pulumi_args="--cwd ${project_dir} --emoji --stack ${PULUMI_STACK}"
 
-cd "${script_dir}/logagent"
-pulumi ${pulumi_args} destroy
+  if [ -f "${project_dir}/Pulumi.yaml" ]; then
+    pulumi ${pulumi_args} destroy
+  else
+    >&2 echo "Not destroying - Pulumi.yaml not found in directory: ${project_dir}"
+  fi
+}
 
-cd "${script_dir}/logstore"
-pulumi ${pulumi_args} destroy
+if command -v aws > /dev/null; then
+  validate_aws_credentials
+fi
 
-cd "${script_dir}/kic-helm-chart"
-pulumi ${pulumi_args} destroy
+k8s_projects=(sirius grafana prometheus certmgr logagent logstore kic-helm-chart)
+if pulumi --cwd "${script_dir}/eks" stack | grep -q 'Current stack resources (0)'; then
+  echo "Pulumi does not know about an EKS instance running"
 
-cd "${script_dir}/kic-image-push"
-pulumi ${pulumi_args} destroy
+  for project in "${k8s_projects[@]}"; do
+    if pulumi --cwd "${script_dir}/${project}" stack | grep -q 'Current stack resources (0)'; then
+      echo "kubernetes project [${project}] has an empty stack - doing nothing"
+    else
+      echo "kubernetes project [${project}] has references in Pulumi - cleaning"
+      stack_name="$(pulumi stack --cwd "${script_dir}/${project}" --show-name)"
+      pulumi stack rm --cwd "${script_dir}/${project}" --force --yes "${stack_name}"
+      pulumi stack init --cwd "${script_dir}/${project}" "${stack_name}"
+      pulumi stack select --cwd "${script_dir}/${project}" "${stack_name}"
+    fi
+  done
+else
+  for project in "${k8s_projects[@]}"; do
+    destroy_project "${project}"
+  done
+fi
 
-cd "${script_dir}/kic-image-build"
-pulumi ${pulumi_args} destroy
+if [[ -n "${1:-}" ]] && [[ "${1}" == "k8s" ]]; then
+  echo "destroyed only kubernetes resources"
+  exit 0
+fi
 
-cd "${script_dir}/ecr"
-pulumi ${pulumi_args} destroy
-
-cd "${script_dir}/eks"
-pulumi ${pulumi_args} destroy
-
-cd "${script_dir}/vpc"
-pulumi ${pulumi_args} destroy
+projects=(kic-image-push kic-image-build ecr eks vpc)
